refactor(arch): consolidate project guard modules (#14351)
