+++ docs/how-tsz-works/file-inventory/README.md
@@ -14,7 +14,7 @@ Then verify coverage with:
 node scripts/docs/check-how-tsz-docs-coverage.mjs
 ```
 
-The latest generation saw 5437 repository files before writing the inventory pages.
+The latest generation saw 5445 repository files before writing the inventory pages.
 
 ## Inventory Pages
 
@@ -24,13 +24,13 @@ The latest generation saw 5437 repository files before writing the inventory pag
 - [Crates Checker Other 01](crates-checker-other-01.md) - mentions 325 files; path: `docs/how-tsz-works/file-inventory/crates-checker-other-01.md`.
 - [Crates Checker Other 02](crates-checker-other-02.md) - mentions 325 files; path: `docs/how-tsz-works/file-inventory/crates-checker-other-02.md`.
 - [Crates Checker Other 03](crates-checker-other-03.md) - mentions 325 files; path: `docs/how-tsz-works/file-inventory/crates-checker-other-03.md`.
-- [Crates Checker Other 04](crates-checker-other-04.md) - mentions 60 files; path: `docs/how-tsz-works/file-inventory/crates-checker-other-04.md`.
+- [Crates Checker Other 04](crates-checker-other-04.md) - mentions 62 files; path: `docs/how-tsz-works/file-inventory/crates-checker-other-04.md`.
 - [Crates Checker Query State Types 01](crates-checker-query-state-types-01.md) - mentions 325 files; path: `docs/how-tsz-works/file-inventory/crates-checker-query-state-types-01.md`.
 - [Crates Checker Query State Types 02](crates-checker-query-state-types-02.md) - mentions 193 files; path: `docs/how-tsz-works/file-inventory/crates-checker-query-state-types-02.md`.
 - [Crates Checker Root G P](crates-checker-root-g-p.md) - mentions 2 files; path: `docs/how-tsz-works/file-inventory/crates-checker-root-g-p.md`.
 - [Crates Checker Root Q Z](crates-checker-root-q-z.md) - mentions 3 files; path: `docs/how-tsz-works/file-inventory/crates-checker-root-q-z.md`.
 - [Crates Checker Test Support 01](crates-checker-test-support-01.md) - mentions 325 files; path: `docs/how-tsz-works/file-inventory/crates-checker-test-support-01.md`.
-- [Crates Checker Test Support 02](crates-checker-test-support-02.md) - mentions 49 files; path: `docs/how-tsz-works/file-inventory/crates-checker-test-support-02.md`.
+- [Crates Checker Test Support 02](crates-checker-test-support-02.md) - mentions 50 files; path: `docs/how-tsz-works/file-inventory/crates-checker-test-support-02.md`.
 - [Crates Core Cli Lsp Wasm Conformance 01](crates-core-cli-lsp-wasm-conformance-01.md) - mentions 325 files; path: `docs/how-tsz-works/file-inventory/crates-core-cli-lsp-wasm-conformance-01.md`.
 - [Crates Core Cli Lsp Wasm Conformance 02](crates-core-cli-lsp-wasm-conformance-02.md) - mentions 325 files; path: `docs/how-tsz-works/file-inventory/crates-core-cli-lsp-wasm-conformance-02.md`.
 - [Crates Core Cli Lsp Wasm Conformance 03](crates-core-cli-lsp-wasm-conformance-03.md) - mentions 294 files; path: `docs/how-tsz-works/file-inventory/crates-core-cli-lsp-wasm-conformance-03.md`.
@@ -39,15 +39,15 @@ The latest generation saw 5437 repository files before writing the inventory pag
 - [Crates Front End And Shared](crates-front-end-and-shared.md) - mentions 269 files; path: `docs/how-tsz-works/file-inventory/crates-front-end-and-shared.md`.
 - [Crates Other](crates-other.md) - mentions 1 file; path: `docs/how-tsz-works/file-inventory/crates-other.md`.
 - [Crates Solver 01](crates-solver-01.md) - mentions 325 files; path: `docs/how-tsz-works/file-inventory/crates-solver-01.md`.
-- [Crates Solver 02](crates-solver-02.md) - mentions 267 files; path: `docs/how-tsz-works/file-inventory/crates-solver-02.md`.
+- [Crates Solver 02](crates-solver-02.md) - mentions 268 files; path: `docs/how-tsz-works/file-inventory/crates-solver-02.md`.
 - [Crates Website](crates-website.md) - mentions 169 files; path: `docs/how-tsz-works/file-inventory/crates-website.md`.
 - [Existing Docs](existing-docs.md) - mentions 39 files; path: `docs/how-tsz-works/file-inventory/existing-docs.md`.
 - [Github And Ci Config](github-and-ci-config.md) - mentions 21 files; path: `docs/how-tsz-works/file-inventory/github-and-ci-config.md`.
-- [How Tsz Works Docs](how-tsz-works-docs.md) - mentions 83 files; path: `docs/how-tsz-works/file-inventory/how-tsz-works-docs.md`.
+- [How Tsz Works Docs](how-tsz-works-docs.md) - mentions 85 files; path: `docs/how-tsz-works/file-inventory/how-tsz-works-docs.md`.
 - [Root And Local Tool Config](root-and-local-tool-config.md) - mentions 12 files; path: `docs/how-tsz-works/file-inventory/root-and-local-tool-config.md`.
-- [Scripts Architecture And Agents](scripts-architecture-and-agents.md) - mentions 33 files; path: `docs/how-tsz-works/file-inventory/scripts-architecture-and-agents.md`.
+- [Scripts Architecture And Agents](scripts-architecture-and-agents.md) - mentions 34 files; path: `docs/how-tsz-works/file-inventory/scripts-architecture-and-agents.md`.
 - [Scripts Bench Ci And Performance](scripts-bench-ci-and-performance.md) - mentions 141 files; path: `docs/how-tsz-works/file-inventory/scripts-bench-ci-and-performance.md`.
 - [Scripts Build Setup Quality And Lsp](scripts-build-setup-quality-and-lsp.md) - mentions 38 files; path: `docs/how-tsz-works/file-inventory/scripts-build-setup-quality-and-lsp.md`.
 - [Scripts Conformance Emit Fourslash](scripts-conformance-emit-fourslash.md) - mentions 67 files; path: `docs/how-tsz-works/file-inventory/scripts-conformance-emit-fourslash.md`.
-- [Scripts Root And Installers](scripts-root-and-installers.md) - mentions 15 files; path: `docs/how-tsz-works/file-inventory/scripts-root-and-installers.md`.
+- [Scripts Root And Installers](scripts-root-and-installers.md) - mentions 16 files; path: `docs/how-tsz-works/file-inventory/scripts-root-and-installers.md`.
 - [File Inventory](README.md) - this index; path: `docs/how-tsz-works/file-inventory/README.md`.
