+++ docs/how-tsz-works/file-inventory/crates-checker-other-01.md
@@ -263,6 +263,7 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `crates/tsz-checker/tests/definite_assignment_tests/part_00.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/definite_assignment_tests/part_01.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/delete_narrowing_reset_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
+| `crates/tsz-checker/tests/destructuring_binding_element_source_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/destructuring_default_flow_assignment_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/destructuring_default_target_narrow_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/destructuring_deferred_conditional_iterability_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
@@ -321,11 +322,10 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `crates/tsz-checker/tests/flow_property_reference_cache_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/for_await_promise_array_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/for_await_type_parameter_iterability_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
+| `crates/tsz-checker/tests/for_in_intersection_operand_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/for_in_narrowing_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/fresh_intersection_display_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/fresh_object_literal_arg_callback_param_variance_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/freshness_stripping_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/fuel_budget_callback_context_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/function_bivariance.rs` | Integration test or fixture for the `tsz-checker` crate. |
-| `crates/tsz-checker/tests/function_expression_excess_param_implicit_any_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
-| `crates/tsz-checker/tests/function_terminal_return_flow_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
