+++ docs/how-tsz-works/file-inventory/crates-checker-other-02.md
@@ -4,6 +4,8 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 
 | Path | Role |
 | --- | --- |
+| `crates/tsz-checker/tests/function_expression_excess_param_implicit_any_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
+| `crates/tsz-checker/tests/function_terminal_return_flow_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/function_type_predicate_alias_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/function_type_predicate_typeof_param_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/generator_annotation_mismatch_display_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
@@ -327,5 +329,3 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `crates/tsz-checker/tests/overload_modifier_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/overload_nontuple_spread_rest_fallback_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/overload_spread_into_rest_ts2556_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
-| `crates/tsz-checker/tests/override_accessor_property_kind_mismatch_ts2416_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
-| `crates/tsz-checker/tests/override_computed_entity_name_ts4127_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
