+++ docs/how-tsz-works/file-inventory/crates-checker-other-03.md
@@ -4,6 +4,8 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 
 | Path | Role |
 | --- | --- |
+| `crates/tsz-checker/tests/override_accessor_property_kind_mismatch_ts2416_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
+| `crates/tsz-checker/tests/override_computed_entity_name_ts4127_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/override_intersection_display_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/parameter_initializer_identifier_context_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/partial_record_optional_index_assignability_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
@@ -327,5 +329,3 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `crates/tsz-checker/tests/type_operator_missing_operand_recovery_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/type_param_modifier_diagnostics_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/type_param_pick_subset_assignability_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
-| `crates/tsz-checker/tests/type_parameter_canonical_typeid_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
-| `crates/tsz-checker/tests/type_parameter_source_constraint_chain_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
