+++ docs/how-tsz-works/file-inventory/crates-checker-other-04.md
@@ -1,9 +1,11 @@
 # Crates Checker Other
 
-Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 60 repository files.
+Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 62 repository files.
 
 | Path | Role |
 | --- | --- |
+| `crates/tsz-checker/tests/type_parameter_canonical_typeid_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
+| `crates/tsz-checker/tests/type_parameter_source_constraint_chain_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/typeof_chained_indexed_access_alias_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/typeof_const_member_conditional_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/tests/typeof_function_generic_alias_narrowing_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
