+++ docs/how-tsz-works/file-inventory/crates-checker-test-support-01.md
@@ -270,6 +270,7 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `crates/tsz-checker/src/tests/overload_literal_source_generalization_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/overload_param_relation_routing_arch_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/overload_two_pass_any_source_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
+| `crates/tsz-checker/src/tests/overload_union_context_callback_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/override_incompatibility_elaboration_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/partial_pick_indexed_access_write_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/polymorphic_this_relation_routing_arch_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
@@ -328,4 +329,3 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `crates/tsz-checker/src/tests/this_prop_nullish_operand_code_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/this_source_inference_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/this_void_method_call_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
-| `crates/tsz-checker/src/tests/truthiness_promise_coercion_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
