+++ docs/how-tsz-works/file-inventory/crates-checker-test-support-02.md
@@ -1,9 +1,10 @@
 # Crates Checker Test Support
 
-Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 49 repository files.
+Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 50 repository files.
 
 | Path | Role |
 | --- | --- |
+| `crates/tsz-checker/src/tests/truthiness_promise_coercion_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/ts1101_with_in_strict_mode_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/ts1170_computed_property_syntactic_form_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
 | `crates/tsz-checker/src/tests/ts18010_jsdoc_tag_anchor_tests.rs` | Integration test or fixture for the `tsz-checker` crate. |
