+++ docs/how-tsz-works/file-inventory/crates-solver-01.md
@@ -217,6 +217,7 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `crates/tsz-solver/src/operations/call_args/string_helpers.rs` | Rust source or test module in the `tsz-solver` crate. |
 | `crates/tsz-solver/src/operations/call_contextual.rs` | Rust source or test module in the `tsz-solver` crate. |
 | `crates/tsz-solver/src/operations/compound_assignment.rs` | Rust source or test module in the `tsz-solver` crate. |
+| `crates/tsz-solver/src/operations/constraints/deferred_conditional.rs` | Rust source or test module in the `tsz-solver` crate. |
 | `crates/tsz-solver/src/operations/constraints/indexed_access.rs` | Rust source or test module in the `tsz-solver` crate. |
 | `crates/tsz-solver/src/operations/constraints/mod.rs` | Rust source or test module in the `tsz-solver` crate. |
 | `crates/tsz-solver/src/operations/constraints/reverse_mapped.rs` | Rust source or test module in the `tsz-solver` crate. |
@@ -328,4 +329,3 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `crates/tsz-solver/src/tests/typedata_contract_tests.rs` | Integration test or fixture for the `tsz-solver` crate. |
 | `crates/tsz-solver/src/tests/visitor_tests.rs` | Integration test or fixture for the `tsz-solver` crate. |
 | `crates/tsz-solver/src/ts_type_flags.rs` | Rust source or test module in the `tsz-solver` crate. |
-| `crates/tsz-solver/src/type_queries/assertion_overlap.rs` | Rust source or test module in the `tsz-solver` crate. |
