+++ docs/how-tsz-works/file-inventory/crates-solver-02.md
@@ -1,9 +1,10 @@
 # Crates Solver
 
-Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 267 repository files.
+Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 268 repository files.
 
 | Path | Role |
 | --- | --- |
+| `crates/tsz-solver/src/type_queries/assertion_overlap.rs` | Rust source or test module in the `tsz-solver` crate. |
 | `crates/tsz-solver/src/type_queries/classifiers.rs` | Rust source or test module in the `tsz-solver` crate. |
 | `crates/tsz-solver/src/type_queries/conditional_infer_alias.rs` | Rust source or test module in the `tsz-solver` crate. |
 | `crates/tsz-solver/src/type_queries/core.rs` | Rust source or test module in the `tsz-solver` crate. |
