+++ docs/how-tsz-works/file-inventory/how-tsz-works-docs.md
@@ -1,6 +1,6 @@
 # How Tsz Works Docs
 
-Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 83 repository files.
+Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 85 repository files.
 
 | Path | Role |
 | --- | --- |
@@ -25,11 +25,13 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 8
 | `docs/how-tsz-works/file-inventory/crates-checker-other-01.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-checker-other-02.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-checker-other-03.md` | Markdown documentation or generated guide page. |
+| `docs/how-tsz-works/file-inventory/crates-checker-other-04.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-checker-query-state-types-01.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-checker-query-state-types-02.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-checker-root-g-p.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-checker-root-q-z.md` | Markdown documentation or generated guide page. |
-| `docs/how-tsz-works/file-inventory/crates-checker-test-support.md` | Markdown documentation or generated guide page. |
+| `docs/how-tsz-works/file-inventory/crates-checker-test-support-01.md` | Markdown documentation or generated guide page. |
+| `docs/how-tsz-works/file-inventory/crates-checker-test-support-02.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-core-cli-lsp-wasm-conformance-01.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-core-cli-lsp-wasm-conformance-02.md` | Markdown documentation or generated guide page. |
 | `docs/how-tsz-works/file-inventory/crates-core-cli-lsp-wasm-conformance-03.md` | Markdown documentation or generated guide page. |
