+++ docs/how-tsz-works/file-inventory/scripts-architecture-and-agents.md
@@ -1,6 +1,6 @@
 # Scripts Architecture And Agents
 
-Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 33 repository files.
+Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 34 repository files.
 
 | Path | Role |
 | --- | --- |
@@ -30,6 +30,7 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `scripts/arch/run-ast-grep.sh` | Architecture guard, ratchet, or source-boundary tooling. |
 | `scripts/arch/test_arch_guard.py` | Architecture guard, ratchet, or source-boundary tooling. |
 | `scripts/arch/test_arch_guard_counts.py` | Architecture guard, ratchet, or source-boundary tooling. |
+| `scripts/arch/test_arch_guard_lsp.py` | Architecture guard, ratchet, or source-boundary tooling. |
 | `scripts/arch/test_arch_guard_misc.py` | Architecture guard, ratchet, or source-boundary tooling. |
 | `scripts/arch/test_arch_guard_policy.py` | Architecture guard, ratchet, or source-boundary tooling. |
 | `scripts/arch/test_arch_guard_project.py` | Architecture guard, ratchet, or source-boundary tooling. |
