+++ docs/how-tsz-works/file-inventory/scripts-root-and-installers.md
@@ -1,6 +1,6 @@
 # Scripts Root And Installers
 
-Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 15 repository files.
+Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 16 repository files.
 
 | Path | Role |
 | --- | --- |
@@ -19,3 +19,4 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 1
 | `scripts/start-website.sh` | Repository automation, tooling, harness, or application source. |
 | `scripts/test/nextest-guard-test.sh` | Repository automation, tooling, harness, or application source. |
 | `scripts/test/nextest-guard.sh` | Repository automation, tooling, harness, or application source. |
+| `scripts/test/safe-run-test.sh` | Repository automation, tooling, harness, or application source. |
